+++ src/include/parser/expression/operator_expression.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include "parser/expression/abstract_expression.h"
 #include "parser/expression_defs.h"
-#include "type/value.h"
+#include "type/transient_value.h"
 
 namespace terrier::parser {
