+++ src/include/type/transient_value_factory.h
@@ -2,7 +2,7 @@
 
 #include <cstring>
 #include "type/type_id.h"
-#include "type/value.h"
+#include "type/transient_value.h"
 
 namespace terrier::type {
 class TransientValueFactory {
