+++ src/include/type/transient_value_peeker.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "type/type_id.h"
-#include "type/value.h"
+#include "type/transient_value.h"
 
 namespace terrier::type {
 class TransientValuePeeker {
